refactor: Stone 클래스와 관련 있는 클래스를 stone 패키지로 위치 변경
