+++ src/main/kotlin/domain/state/BlackTurn.kt
@@ -1,6 +1,6 @@
 package domain.state
 
-import domain.Stone
+import domain.stone.Stone
 
 class BlackTurn(override val blackStones: Set<Stone>, override val whiteStones: Set<Stone>) :
     Running(blackStones, whiteStones) {
