+++ src/main/kotlin/domain/state/BlackWin.kt
@@ -1,6 +1,6 @@
 package domain.state
 
-import domain.Stone
+import domain.stone.Stone
 
 class BlackWin(override val blackStones: Set<Stone>, override val whiteStones: Set<Stone>) :
     Finished(blackStones, whiteStones) {
