+++ src/main/kotlin/domain/state/WhiteTurn.kt
@@ -1,6 +1,6 @@
 package domain.state
 
-import domain.Stone
+import domain.stone.Stone
 
 class WhiteTurn(override val blackStones: Set<Stone>, override val whiteStones: Set<Stone>) :
     Running(blackStones, whiteStones) {
