+++ src/main/kotlin/domain/state/WhiteWin.kt
@@ -1,6 +1,6 @@
 package domain.state
 
-import domain.Stone
+import domain.stone.Stone
 
 class WhiteWin(override val blackStones: Set<Stone>, override val whiteStones: Set<Stone>) :
     Finished(blackStones, whiteStones) {
