+++ src/main/kotlin/view/BoardView.kt
@@ -1,9 +1,9 @@
 package view
 
 import domain.OmokGame
-import domain.Stone
-import domain.XCoordinate
-import domain.YCoordinate
+import domain.stone.Stone
+import domain.stone.XCoordinate
+import domain.stone.YCoordinate
 
 class BoardView(omokGame: OmokGame) {
