+++ src/test/kotlin/domain/YCoordinateTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.stone.YCoordinate
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
