+++ src/test/kotlin/domain/rule/BlackStonesAreNotAllowed44RuleTest.kt
@@ -1,8 +1,8 @@
 package domain.rule
 
-import domain.Stone
 import domain.state.BlackTurn
 import domain.state.State
+import domain.stone.Stone
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.TestInstance
